refactor: modify hook deps
