+++ src/hooks/usePdfScore.ts
